feat: Add API endpoint to retrieve product image
The `product_image.php` file has been added to the API directory. This file contains code to retrieve and display a product image based on the provided product ID and image ID. The image quality can be adjusted using the `quality` parameter, with a default value of 50. This API endpoint will be used to fetch product images for display on the website.