Refactor products.php to include category filter in SQL query
The `products.php` file has been refactored to include a category filter in the SQL query. If the `category` parameter is provided in the GET request, the query will filter the products based on the specified category ID. This change allows for more targeted product retrieval and display on the website.